% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_invoice.R
\name{create_invoice}
\alias{create_invoice}
\title{Create a new document based on the \code{invoicer} template}
\usage{
create_invoice(file, create_dir = FALSE, edit = FALSE)
}
\arguments{
\item{file}{File name for the draft}

\item{create_dir}{TRUE to create a new directory for the draft}

\item{edit}{TRUE to edit the template immediately}
}
\value{
The contents of the \code{invoicer} template produced at specified \code{file}
}
\description{
Create (and optionally edit) a draft of the \code{invoicer} R Markdown template.
}
