% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape_latex.R
\name{escape_latex}
\alias{escape_latex}
\title{Escape problematic characters to allow compilation in LaTeX}
\usage{
escape_latex(x, newlines = FALSE, spaces = FALSE)
}
\arguments{
\item{x}{a character string to escape}

\item{newlines}{boolean}

\item{spaces}{boolean}
}
\description{
This function escapes special LaTeX characters in a string to ensure proper rendering.
Copied from \code{knitr} because it is unexported and CRAN rejects :::
}
\examples{
escape_latex("100\% sure!")
}
